refactor(auth): modularize AuthHandler and bump dependencies
- Move configuration logic to ConfigService
- Move route definitions to separate file
- Update API entry point to use AuthHandler.routes property
- Bump turbo to 2.9.6 and bun-types to 1.3.13
- Add version field to @dockstat/auth package.json
